fix: Fixed type StockDividend price

+++ src/main/java/com/nfinance/quotes/StockDividend.java
@@ -6,5 +6,5 @@ public class StockDividend {
 
     private Double yieldRate;
 
-    private Double yieldPrice;
+    private BigDecimal yieldPrice;
 }
